feat: add dashboard-as-code schema generator and regenerate schemas
Extract shared utilities from the chart schema generator into
schemaUtils.ts, then add a dashboard schema generator that derives
dashboard-as-code-1.0.json from the swagger spec. This fixes the
schema drift that caused lint failures (missing hideFrame, content
on markdown tiles).

Both schemas are now generated from the TSOA swagger output,
making TypeScript types the single source of truth.
